move definition out of __init__
